+++ src/main/java/org/openlmis/requisition/service/referencedata/SupervisoryNodeReferenceDataService.java
@@ -19,7 +19,9 @@ import org.openlmis.requisition.dto.SupervisoryNodeDto;
 import org.openlmis.requisition.service.RequestParameters;
 import org.springframework.stereotype.Service;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 
 @Service
@@ -45,13 +47,13 @@ public class SupervisoryNodeReferenceDataService
    * Find a correct supervisory node by the provided facility and program.
    */
   public SupervisoryNodeDto findSupervisoryNode(UUID program, UUID facility) {
-    RequestParameters parameters = RequestParameters
-        .init()
-        .set("programId", program)
-        .set("facilityId", facility);
+    Map<String, Object> requestBody = new HashMap<>();
+    requestBody.put("programId", program);
+    requestBody.put("facilityId", facility);
 
-    List<SupervisoryNodeDto> supervisoryNodeDtos = findAll("search", parameters);
-    return supervisoryNodeDtos.isEmpty() ? null : supervisoryNodeDtos.get(0);
+    List<SupervisoryNodeDto> content =
+        getPage("search", RequestParameters.init(), requestBody).getContent();
+    return content.size() > 0 ? content.get(0) : null;
   }
 
 }